Move publishing into its own script
