+++ core/build.gradle.kts
@@ -1,12 +1,12 @@
 plugins {
     `lavalink-module`
+    `lavalink-publishing`
     kotlin("plugin.serialization")
     id("org.jetbrains.dokka")
     id("kotlinx-atomicfu")
 }
 
 kotlin {
-
     sourceSets {
         commonMain {
             dependencies {
@@ -63,5 +63,3 @@ kotlin {
         }
     }
 }
-
-applyPublishing()
